trying out

+++ src/main/java/com/abm/moviesapi/controller/MoviesRestController.java
@@ -1,6 +1,7 @@
 package com.abm.moviesapi.controller;
 
 import com.abm.moviesapi.entity.Movie;
+import com.abm.moviesapi.exceptions.CustomMovieException.MovieNotFoundException;
 import com.abm.moviesapi.service.MovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -27,35 +28,37 @@ public class MoviesRestController {
     //GET (movie by id)
     @GetMapping("/movies/{movieId}")
     public Movie getMovie(@PathVariable int movieId) {
-        return movieService.findById(movieId);
+        Movie movie = movieService.findById(movieId);
+        if (movie.getId() < 0 || movie.getId() == 0){
+            throw new MovieNotFoundException();
+        }
+        return movie;
     }
 
     //POST - add new one
     @PostMapping(value = "/movies", consumes = MediaType.APPLICATION_JSON_VALUE)
     public Movie addMovie(@RequestBody Movie movie){
-        try {
-            movie.setId(0); // if it isn't here spring will be trying to update element
-            movieService.save(movie);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        movie.setId(0); // if it isn't here spring will be trying to update element
+        movieService.save(movie);
         return movie;
     }
 
-    //PUT update existing one
     @PutMapping("/movies")
     public Movie updateMovie(@RequestBody Movie movie){
-        try {
-            movieService.save(movie);
-        } catch (Exception e) {
-            e.printStackTrace();
+        if (movie.getId() <= 0){
+            throw new MovieNotFoundException();
         }
+        movieService.save(movie);
         return movie;
     }
     //DELETE
     @DeleteMapping("/movies/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public HttpStatus deleteMovie(@PathVariable int id){
+        Movie movie = movieService.findById(id);
+        if (movie.getId()<=0){
+            throw new MovieNotFoundException();
+        }
         movieService.deleteById(id);
         return HttpStatus.NO_CONTENT;
     }
